refactor: move query building logic, add automatic semantic scoring
