+++ tests/executor/indexer/base/test_base_executor.py
@@ -1,26 +1,34 @@
 import os
+import time
 from copy import deepcopy
 
 import numpy as np
 import pytest
+from elasticsearch import Elasticsearch
 from jina import Document, DocumentArray, Flow
 
 from now.constants import TAG_INDEXER_DOC_HAS_TEXT, TAG_OCR_DETECTOR_TEXT_IN_DOC
 from now.deployment.deployment import cmd
+from now.executor.indexer.elastic import NOWElasticIndexer
 from now.executor.indexer.in_memory.in_memory_indexer import InMemoryIndexer
 from now.executor.indexer.qdrant import NOWQdrantIndexer16
 
 NUMBER_OF_DOCS = 10
 DIM = 128
+MAX_RETRIES = 20
 
 
 @pytest.mark.parametrize(
-    'indexer',
-    [InMemoryIndexer, NOWQdrantIndexer16],
+    'indexer,setup',
+    [
+        (InMemoryIndexer, None),
+        (NOWQdrantIndexer16, 'setup_qdrant'),
+        (NOWElasticIndexer, 'setup_elastic'),
+    ],
 )
 class TestBaseIndexer:
-    @pytest.fixture(scope='function', autouse=True)
-    def setup(self):
+    @pytest.fixture(scope='function')
+    def setup_qdrant(self):
         cur_dir = os.path.dirname(os.path.abspath(__file__))
         compose_yml = os.path.abspath(os.path.join(cur_dir, 'docker-compose.yml'))
 
@@ -32,9 +40,25 @@ class TestBaseIndexer:
             f"docker-compose -f {compose_yml} --project-directory . down --remove-orphans"
         )
 
-    @pytest.fixture(scope='function', autouse=True)
-    def metas(self, tmpdir):
-        return {'workspace': str(tmpdir)}
+    @pytest.fixture(scope='function')
+    def setup_elastic(self):
+        hosts = 'http://localhost:9200'
+        cmd('docker-compose -f tests/resources/elastic/docker-compose.yml up -d')
+        retries = 0
+        while retries < MAX_RETRIES:
+            try:
+                es = Elasticsearch(hosts=hosts)
+                if es.ping():
+                    break
+                else:
+                    retries += 1
+                    time.sleep(5)
+            except Exception:
+                print('Elasticsearch is not running')
+        if retries >= MAX_RETRIES:
+            raise RuntimeError('Elasticsearch is not running')
+        yield
+        cmd('docker-compose -f tests/resources/elastic/docker-compose.yml down')
 
     def gen_docs(self, num):
         res = DocumentArray()
@@ -85,8 +109,10 @@ class TestBaseIndexer:
 
         return da
 
-    def test_index(self, tmpdir, indexer):
+    def test_index(self, tmpdir, indexer, setup, request):
         """Test indexing does not return anything"""
+        if setup:
+            request.getfixturevalue(setup)
         metas = {'workspace': str(tmpdir)}
         docs = self.gen_docs(NUMBER_OF_DOCS)
         f = Flow().add(
@@ -103,8 +129,9 @@ class TestBaseIndexer:
     @pytest.mark.parametrize(
         'offset, limit', [(0, 10), (10, 0), (0, 0), (10, 10), (None, None)]
     )
-    def test_list(self, offset, limit, indexer, metas):
+    def test_list(self, tmpdir, offset, limit, indexer, setup):
         """Test list returns all indexed docs"""
+        metas = {'workspace': str(tmpdir)}
         docs = self.gen_docs(NUMBER_OF_DOCS)
         f = Flow().add(
             uses=indexer,
@@ -132,7 +159,8 @@ class TestBaseIndexer:
                 assert [d.uri for d in list_res] == ['my-parent-uri'] * l
                 assert [d.tags['parent_tag'] for d in list_res] == ['value'] * l
 
-    def test_search(self, indexer, metas):
+    def test_search(self, tmpdir, indexer, setup):
+        metas = {'workspace': str(tmpdir)}
         docs = self.gen_docs(NUMBER_OF_DOCS)
         docs_query = self.gen_docs(1)
         f = Flow().add(
@@ -154,7 +182,8 @@ class TestBaseIndexer:
                     <= query_res[0].matches[i + 1].scores['cosine'].value
                 )
 
-    def test_search_match(self, indexer, metas):
+    def test_search_match(self, tmpdir, indexer, setup):
+        metas = {'workspace': str(tmpdir)}
         docs = self.gen_docs(NUMBER_OF_DOCS)
         docs_query = self.gen_docs(NUMBER_OF_DOCS)
         f = Flow().add(
@@ -184,7 +213,8 @@ class TestBaseIndexer:
                     <= c.matches[i + 1].scores['cosine'].value
                 )
 
-    def test_search_with_filtering(self, indexer, metas):
+    def test_search_with_filtering(self, tmpdir, indexer, setup):
+        metas = {'workspace': str(tmpdir)}
 
         docs = self.docs_with_tags(NUMBER_OF_DOCS)
         docs_query = self.gen_docs(1)
@@ -205,7 +235,8 @@ class TestBaseIndexer:
             )
             assert all([m.tags['price'] < 50 for m in query_res[0].matches])
 
-    def test_delete(self, indexer, metas):
+    def test_delete(self, tmpdir, indexer, setup):
+        metas = {'workspace': str(tmpdir)}
         docs = self.gen_docs(NUMBER_OF_DOCS)
         f = Flow().add(
             uses=indexer,
@@ -228,7 +259,8 @@ class TestBaseIndexer:
             docs_query = self.gen_docs(NUMBER_OF_DOCS)
             f.post(on='/search', inputs=docs_query, return_results=True)
 
-    def test_get_tags(self, indexer, metas):
+    def test_get_tags(self, tmpdir, indexer, setup):
+        metas = {'workspace': str(tmpdir)}
         docs = DocumentArray(
             [
                 Document(
@@ -276,7 +308,8 @@ class TestBaseIndexer:
                 0
             ].tags['tags']['color'] == ['blue', 'red']
 
-    def test_delete_tags(self, indexer, metas):
+    def test_delete_tags(self, tmpdir, indexer, setup):
+        metas = {'workspace': str(tmpdir)}
         docs = DocumentArray(
             [
                 Document(
@@ -443,8 +476,9 @@ class TestBaseIndexer:
         ],
     )
     def test_search_chunk_using_sum_ranker(
-        self, documents, indexer, query, embedding, res_ids, metas
+        self, documents, indexer, setup, query, embedding, res_ids, tmpdir
     ):
+        metas = {'workspace': str(tmpdir)}
         documents = DocumentArray([Document(chunks=[doc]) for doc in documents])
         with Flow().add(
             uses=indexer,
@@ -475,7 +509,8 @@ class TestBaseIndexer:
                 if d.uri:
                     assert d.blob == b'', f'got blob {d.blob} for {d.id}'
 
-    def test_no_blob_with_working_uri(self, indexer, metas):
+    def test_no_blob_with_working_uri(self, tmpdir, indexer, setup):
+        metas = {'workspace': str(tmpdir)}
         with Flow().add(
             uses=indexer,
             uses_with={
@@ -526,9 +561,9 @@ class TestBaseIndexer:
             assert matches[3].blob == b''
             assert matches[4].tensor is None
 
-    def test_curate_endpoint(self, indexer, metas):
+    def test_curate_endpoint(self, tmpdir, indexer, setup):
         """Test indexing does not return anything"""
-
+        metas = {'workspace': str(tmpdir)}
         docs = self.gen_docs(NUMBER_OF_DOCS)
         docs.append(
             Document(
@@ -560,7 +595,6 @@ class TestBaseIndexer:
             uses_metas=metas,
         )
         with f:
-            f.index(docs, return_results=True)
             f.post(
                 on='/curate',
                 parameters={
@@ -572,6 +606,7 @@ class TestBaseIndexer:
                     }
                 },
             )
+            f.index(docs, return_results=True)
             result = f.search(
                 inputs=Document(
                     chunks=[
@@ -604,18 +639,3 @@ class TestBaseIndexer:
                     ]
                 )
             )
-
-    def test_curate_endpoint_incorrect(self, indexer, metas):
-        f = Flow().add(
-            uses=indexer,
-            uses_with={
-                'dim': DIM,
-            },
-            uses_metas=metas,
-        )
-        with f:
-            with pytest.raises(Exception):
-                f.post(
-                    on='/curate',
-                    parameters={'queryfilter': {}},
-                )
